Corrigido Builder::delTree() para não dar erro se tentar apagar um diretórios que não existe.

+++ src/Build/Builder.php
@@ -185,6 +185,8 @@ class Builder
      */
     public static function delTree(string $dir)
     {
+        if(!file_exists($dir)) return;
+        
         $files = array_diff(scandir($dir), array('.', '..'));
         foreach ($files as $file) {
             (is_dir("$dir/$file")) ? self::delTree("$dir/$file") : unlink("$dir/$file");
